requirements.txt
aiogram~=3.0.0b7
environs~=9.5.0
Pillow~=9.4.0